fix: add checking for user existance in users service

+++ backend/src/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -9,7 +9,7 @@ export class UsersService {
 	constructor(private readonly prisma: PrismaService) {}
 
 	async findById(id: number): Promise<User | null> {
-		return await this.prisma.user.findUnique({
+		return this.prisma.user.findUnique({
 			where: {
 				id: id,
 			},
@@ -17,7 +17,7 @@ export class UsersService {
 	}
 
 	async findByLogin(login: string): Promise<User | null> {
-		return await this.prisma.user.findUnique({
+		return this.prisma.user.findUnique({
 			where: {
 				login: login,
 			},
@@ -25,7 +25,7 @@ export class UsersService {
 	}
 
 	async findByEmail(email: string): Promise<User | null> {
-		return await this.prisma.user.findUnique({
+		return this.prisma.user.findUnique({
 			where: {
 				email: email,
 			},
@@ -33,7 +33,7 @@ export class UsersService {
 	}
 
 	async findByVerifyToken(token: string): Promise<User | null> {
-		return await this.prisma.user.findUnique({
+		return this.prisma.user.findUnique({
 			where: {
 				verifyToken: token,
 			},
@@ -45,12 +45,24 @@ export class UsersService {
 	}
 
 	async create(dto: CreateUserDto): Promise<User> {
+		const { login, email } = dto;
+
+		await this.checkIfNotExist(login, email);
+
 		return this.prisma.user.create({
 			data: dto,
 		});
 	}
 
 	async update(id: number, dto: UpdateUserDto): Promise<User> {
+		const user = await this.findById(id);
+
+		if (!user) {
+			throw new BadRequestException(`User with id ${id} doesn't exist`);
+		}
+
+		await this.checkIfNotExist(dto.login, dto.email);
+
 		return this.prisma.user.update({
 			where: {
 				id: id,
@@ -60,6 +72,12 @@ export class UsersService {
 	}
 
 	async delete(id: number): Promise<User> {
+		const user = await this.findById(id);
+
+		if (!user) {
+			throw new BadRequestException(`User with id ${id} doesn't exist`);
+		}
+
 		return this.prisma.user.delete({
 			where: {
 				id: id,
@@ -68,6 +86,12 @@ export class UsersService {
 	}
 
 	async setAvatar(id: number, path: string): Promise<User> {
+		const user = await this.findById(id);
+
+		if (!user) {
+			throw new BadRequestException(`User with id ${id} doesn't exist`);
+		}
+
 		return this.prisma.user.update({
 			where: {
 				id: id,
@@ -76,5 +100,25 @@ export class UsersService {
 				avatar: path,
 			},
 		});
+	}
+
+	private async checkIfNotExist(login: string, email: string): Promise<void> {
+		if (login) {
+			const userByLogin = await this.findByLogin(login);
+			if (userByLogin) {
+				throw new BadRequestException(
+					`User with login ${login} already exists`,
+				);
+			}
+		}
+
+		if (email) {
+			const userByEmail = await this.findByEmail(email);
+			if (userByEmail) {
+				throw new BadRequestException(
+					`User with email ${email} already exists`,
+				);
+			}
+		}
 	}
 }
